Solve minor issues
