+++ pkg/rpc/doc.go
@@ -1,2 +1,2 @@
-// TODO: Document package.
+// Package rpc represent all the set of instructions related to rpc calls, over JSON-RPC 2.0
 package rpc
