+++ pkg/rpc/types.go
@@ -2,16 +2,16 @@ package rpc
 
 import "github.com/NethermindEth/juno/pkg/types"
 
-// TODO: Document.
+// EventRequest represent allOf types.EventFilter and types.ResultPageRequest
 type EventRequest struct {
 	types.EventFilter
 	types.ResultPageRequest
 }
 
-// TODO: Document.
+// EmittedEventArray represent a set of emmited events
 type EmittedEventArray []types.EmittedEvent
 
-// TODO: Document.
+// EventResponse represent the struct of the response of events
 type EventResponse struct {
 	EmittedEventArray
 	PageNumber uint64 `json:"page_number"`
