--- pkg/rpc/utils.go
@@ -1,16 +1,0 @@
-package rpc
-
-import (
-	"encoding/json"
-	"fmt"
-)
-
-// TODO: Document.
-func StructPrinter(i interface{}) {
-	// notest
-	b, err := json.Marshal(i)
-	if err != nil {
-		return
-	}
-	fmt.Println(string(b))
-}
